requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.1
django-extensions==3.2.3
django-heroku==0.3.1
et-xmlfile==1.1.0
gunicorn==21.2.0
lxml==5.2.1
openpyxl==3.1.2
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-docx==1.1.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.9.0
tzdata==2023.4
waitress==3.0.0
whitenoise==6.6.0
